feat: seperation of request and data models
