+++ BittrexTracker/DataModels/Coins.swift
@@ -8,19 +8,6 @@
 
 import Foundation
 
-/// Coin Request struct to hold request state and result
-struct CoinRequest: Decodable {
-  let success: Bool?
-  let message: String?
-  let result: [Coin]?
-  
-  init(success: Bool?, message: String?, result: [Coin]?) {
-    self.success = success
-    self.message = message
-    self.result = result
-  }
-}
-
 /// The data structure for the result of the Coin Request
 struct Coin: Decodable {
   let currency: String
