+++ BittrexTracker/DataModels/DepositHistory.swift
@@ -1,5 +1,5 @@
 //
-//  WithdrawalHistoryRequest.swift
+//  DepositHistoryRequest.swift
 //  BittrexTracker
 //
 //  Created by Matthew Dovey on 21/01/2019.
@@ -8,28 +8,15 @@
 
 import Foundation
 
-/// Withdrawal History Request struct to hold request state and result
-struct WithdrawalHistoryRequest: Decodable {
-  let success: Bool?
-  let message: String?
-  let result: [WithdrawalHistory]?
-  
-  init(success: Bool?, message: String?, result: [WithdrawalHistory]?) {
-    self.success = success
-    self.message = message
-    self.result = result
-  }
-}
-
-/// The data structure for the result of the Withdrawal History Request
-struct WithdrawalHistory: Decodable {
+/// The data structure for the result of the Deposit History Request
+struct DepositHistory: Decodable {
   let paymentUuid: String?
   let currency: String?
   let amount: Double?
   let address: String?
   let opened: String?
   let authorized: Bool?
-  let pendingPayments: Bool?
+  let pendingPayment: Bool?
   let txCost: Double?
   let txId: String?
   let canceled: Bool?
@@ -42,7 +29,7 @@ struct WithdrawalHistory: Decodable {
     case address = "Address"
     case opened = "Opened"
     case authorized = "Authorized"
-    case pendingPayments = "PendingPayments"
+    case pendingPayment = "PendingPayment"
     case txCost = "TxCost"
     case txId = "TxID"
     case canceled = "Canceled"
