+++ BittrexTracker/DataModels/MarketHistory.swift
@@ -8,19 +8,6 @@
 
 import Foundation
 
-/// Market History Request struct to hold request state and result
-struct MarketHistoryRequest: Decodable {
-  let success: Bool?
-  let message: String?
-  let result: [MarketHistory]?
-  
-  init(success: Bool?, message: String?, result: [MarketHistory]?) {
-    self.success = success
-    self.message = message
-    self.result = result
-  }
-}
-
 /// The data structure for the result of the Market History Request
 struct MarketHistory: Decodable {
   let id: String?
