+++ BittrexTracker/DataModels/MarketSummary.swift
@@ -8,19 +8,6 @@
 
 import Foundation
 
-/// Market Summary Request struct to hold request state and result
-struct MarketSummaryRequest: Decodable {
-  let success: Bool?
-  let message: String?
-  let result: [MarketSummary]?
-  
-  init(success: Bool?, message: String?, result: [MarketSummary]?) {
-    self.success = success
-    self.message = message
-    self.result = result
-  }
-}
-
 /// The data structure for the result of the Market Summary Request
 struct MarketSummary: Decodable {
   let marketName: String
