+++ BittrexTracker/DataModels/OrderHistory.swift
@@ -8,19 +8,6 @@
 
 import Foundation
 
-/// Order History Request struct to hold request state and result
-struct OrderHistoryRequest: Decodable {
-  let success: Bool?
-  let message: String?
-  let result: [OrderHistory]?
-  
-  init(success: Bool?, message: String?, result: [OrderHistory]?) {
-    self.success = success
-    self.message = message
-    self.result = result
-  }
-}
-
 /// The data structure for the result of the Order History Request
 struct OrderHistory: Decodable {
   let orderUuid: String?
