+++ BittrexTracker/RequestModels/CoinRequest.swift
@@ -1,0 +1,22 @@
+//
+//  CoinRequest.swift
+//  BittrexTracker
+//
+//  Created by Matthew Dovey on 29/01/2019.
+//  Copyright © 2019 Matthew Dovey. All rights reserved.
+//
+
+import Foundation
+
+/// Coin Request struct to hold request state and result
+struct CoinRequest: Decodable {
+  let success: Bool?
+  let message: String?
+  let result: [Coin]?
+  
+  init(success: Bool?, message: String?, result: [Coin]?) {
+    self.success = success
+    self.message = message
+    self.result = result
+  }
+}
