+++ BittrexTracker/RequestModels/MarketHistoryRequest.swift
@@ -1,5 +1,5 @@
 //
-//  Balance.swift
+//  MarketHistoryRequest.swift
 //  BittrexTracker
 //
 //  Created by Matthew Dovey on 29/01/2019.
@@ -8,13 +8,13 @@
 
 import Foundation
 
-/// Balance Request struct to hold request state and result
-struct BalanceRequest: Decodable {
+/// Market History Request struct to hold request state and result
+struct MarketHistoryRequest: Decodable {
   let success: Bool?
   let message: String?
-  let result: Balance?
+  let result: [MarketHistory]?
   
-  init(success: Bool?, message: String?, result: Balance?) {
+  init(success: Bool?, message: String?, result: [MarketHistory]?) {
     self.success = success
     self.message = message
     self.result = result
